+++ .settings/org.eclipse.core.resources.prefs
@@ -1,6 +1,6 @@
 eclipse.preferences.version=1
 encoding//src/main/java=UTF-8
 encoding//src/main/resources=UTF-8
-encoding//src/test/java=UTF-8
-encoding//src/test/resources=UTF-8
+encoding//src/testQuotes/java=UTF-8
+encoding//src/testQuotes/resources=UTF-8
 encoding/<project>=UTF-8
