+++ src/test/java/testQuotes/QuoteServiceTest.java
@@ -1,4 +1,4 @@
-package test;
+package testQuotes;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
